% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.contrast.R
\name{print.contrast}
\alias{print.contrast}
\title{Print a Contrast Object}
\usage{
\method{print}{contrast}(x, X = FALSE, fun = function(u) u, ...)
}
\arguments{
\item{x}{Result of \code{contrast()}.}

\item{X}{A logical: set \code{TRUE} to print design matrix used in computing the
contrasts (or the average contrast).}

\item{fun}{A function to transform the contrast, SE, and lower and upper
confidence limits before printing.  For example, specify \code{fun = exp} to
anti-log them for logistic models.}

\item{...}{Not used.}
}
\description{
Print a Contrast Object
}
