% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{two_factor_crossed}
\alias{two_factor_crossed}
\title{Complete Two-Factor Experiment}
\value{
\item{two_factor_crossed}{A data frame}
}
\description{
Complete Two-Factor Experiment
}
\details{
A gene expression experiment was run to assess the effect of a compound
under two different diets: high fat and low fat. The main comparisons of
interest are the difference between the treated and untreated groups within
a diet. The interaction effect was a secondary hypothesis. For illustration,
we only include the expression value of one of the genes. The study design
was a full two-way factorial with \code{n = 24} samples.
}
\examples{
two_factor_crossed
}
\keyword{datasets}
